add Chart component
